rewrite virtual network layer code but no test.

+++ trunk/p2p_slotion/mediator_pattern.h
@@ -131,8 +131,8 @@ public:
   }
 public:
   //application layer
-  sigslot::signal4<int ,SocketType,const char *, int> SignalSendDataToUpLayer;
-  virtual void OnReceiveDataFromUpLayer(int,SocketType,const char *, int) = 0;
+  sigslot::signal4<uint32 ,SocketType,const char *, uint16> SignalSendDataToUpLayer;
+  virtual void OnReceiveDataFromUpLayer(uint32,SocketType,const char*,uint16) = 0;
 public:
   //ice part 
   sigslot::signal2<const char *, int> SignalSendDataToLowLayer;
@@ -164,9 +164,9 @@ public:
   //  virtual void BindTcpPort(int tcp_port) = 0;
   //  virtual void BindUdpPort(int udp_port) = 0;
 public: // virtual network interface
-  sigslot::signal4<int ,SocketType,const char *, int> SignalSendDataToLowLayer;
-  virtual void OnReceiveDateFromLowLayer(int socket, SocketType socket_type,
-    const char *data, int len) = 0;
+  sigslot::signal4<uint32,SocketType,const char*,uint16> SignalSendDataToLowLayer;
+  virtual void OnReceiveDateFromLowLayer(uint32 socket, SocketType socket_type,
+    const char *data, uint16 len) = 0;
 private:
   AbstractVirtualNetwork  *virtual_network_;
   DISALLOW_EVIL_CONSTRUCTORS(AbstarctVirtualApplication);
